Added AddMigration Commands, DBContext Modules And etc...
