+++ Infrastructure/Common/ModuleDbContext.cs
@@ -1,0 +1,23 @@
+using Microsoft.EntityFrameworkCore;
+
+namespace Infrastructure.Common;
+
+public abstract class ModuleDbContext : DbContext
+{
+    protected abstract string Schema { get; }
+
+    protected ModuleDbContext(DbContextOptions options) : base(options)
+    {
+    }
+
+    protected override void OnModelCreating(ModelBuilder modelBuilder)
+    {
+        if (!string.IsNullOrWhiteSpace(Schema))
+            modelBuilder.HasDefaultSchema(Schema);
+
+        base.OnModelCreating(modelBuilder);
+        modelBuilder.ApplyConfigurationsFromAssembly(GetType().Assembly);
+    }
+
+    public override async Task<int> SaveChangesAsync(CancellationToken cancellationToken = default) => await base.SaveChangesAsync(true, cancellationToken);
+}
