+++ Infrastructure/Persistence/Configurations/PromotionCategoriesConfigurations.cs
