+++ Infrastructure/Persistence/Configurations/PromotionProductsConfigurations.cs
